+++ src/main/java/se/kth/bbc/jobs/adam/AdamJob.java
@@ -5,9 +5,7 @@ import java.util.logging.Level;
 import java.util.logging.Logger;
 import se.kth.bbc.fileoperations.FileOperations;
 import se.kth.bbc.jobs.jobhistory.ExecutionFacade;
-import se.kth.bbc.jobs.jobhistory.JobOutputFile;
 import se.kth.bbc.jobs.jobhistory.JobOutputFileFacade;
-import se.kth.bbc.jobs.jobhistory.JobOutputFilePK;
 import se.kth.bbc.jobs.yarn.YarnJob;
 import se.kth.bbc.jobs.yarn.YarnRunner;
 import se.kth.bbc.lims.Utils;
